features(quality) flake8 improve
