+++ .github/workflows/pythonapp.yml
@@ -23,8 +23,8 @@ jobs:
         pip install flake8
         # stop the build if there are Python syntax errors or undefined names
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+        # The GitHub editor is 127 chars wide, but pycharm is not :D
+        flake8 . --count --max-complexity=10 --max-line-length=150 --statistics --ignore=E401,W605
     - name: Run coverage
       run: |
         coverage run --source='.' --rcfile=.coveragerc src/manage.py test  -v 2
