+++ Dockerfile
@@ -1,6 +1,9 @@
 FROM bitnami/python:3.7
 
-RUN apt-get update && apt-get install redis-server sqlite3 -y
+RUN apt-get update \
+ && apt-get install redis-server sqlite3 -y --no-install-recommends \
+ && apt-get clean \
+ && rm -rf /var/lib/apt/lists/*
 
 RUN groupadd -g 1001 app && useradd -r -u 1001 -g app app
 RUN mkdir /home/app && chown 1001 /home/app
