+++ src/sensors/sensors3_linechart.py
@@ -1,6 +1,6 @@
-import requests, time
-from src.tipboard.app.properties import TIPBOARD_URL
-from src.sensors.utils import end, buildConfigTile
+import time
+from src.sensors.utils import end
+from src.sensors.utils import sendDataToTipboard
 
 NAME_OF_SENSORS = "linechart"
 TILE_TEMPLATE = "line_chart"
@@ -13,16 +13,11 @@ def executeScriptToGetData():
              ["27.09", 190240], ["28.09", 125272], ["29.09", 3685]]
     line2 = [["23.09", 3685], ["24.09", 125272], ["25.09", 190240], ["26.09", 229639],
              ["27.09", 240933], ["28.09", 260630], ["29.09", 108326]]
-    return {"subtitle": "averages from last week",
-               "description": "Sales in our dept",
-               "series_list": [line1, line2]}
-
-
-def sendDataToTipboard(data=None, tile_template=None, tile_id="", isTest=False):
-    configTile = buildConfigTile(tile_id=tile_id, tile_template=tile_template, data=data)
-    if not isTest:
-        res = requests.post(TIPBOARD_URL + "/push", data=configTile)
-        print(f"{res} -> {tile_id}: {res.text}", flush=True)
+    return {
+        "subtitle": "averages from last week",
+        "description": "Sales in our dept",
+        "series_list": [line1, line2]
+    }
 
 
 def sonde3(isTest):
