+++ src/sensors/sensors5_simplepercentage.py
@@ -1,6 +1,6 @@
-import requests, time
-from src.tipboard.app.properties import TIPBOARD_URL
-from src.sensors.utils import end, buildConfigTile
+import time
+from src.sensors.utils import end
+from src.sensors.utils import sendDataToTipboard
 
 NAME_OF_SENSORS = "simple_percentage"
 TILE_TEMPLATE = "simple_percentage"
@@ -12,15 +12,8 @@ def executeScriptToGetData():
     return {"title": "My title",
             "subtitle": "My subtitle",
             "big_value": "100%",
-            "left_label":  "smaller label 1", "left_value": "50%",
+            "left_label": "smaller label 1", "left_value": "50%",
             "right_label": "smaller label 2", "right_value": "25%"}
-
-
-def sendDataToTipboard(data=None, tile_template=None, tile_id="", isTest=False):
-    configTile = buildConfigTile(tile_id=tile_id, tile_template=tile_template, data=data)
-    if not isTest:
-        res = requests.post(TIPBOARD_URL + "/push", data=configTile)
-        print(f"{res} -> {tile_id}: {res.text}", flush=True)
 
 
 def sonde5(isTest):
