+++ src/sensors/sensors6_listing.py
@@ -1,6 +1,6 @@
-import requests, time
-from src.tipboard.app.properties import TIPBOARD_URL
-from src.sensors.utils import end, buildConfigTile
+import time
+from src.sensors.utils import end
+from src.sensors.utils import sendDataToTipboard
 
 NAME_OF_SENSORS = "listing"
 TILE_TEMPLATE = "listing"
@@ -14,15 +14,9 @@ def executeScriptToGetData():
             ["Leader: 5",
              "Product Owner: 0",
              "Scrum Master: 3",
-             "Developer: 0"]
-        }
-
-
-def sendDataToTipboard(data=None, tile_template=None, tile_id="", isTest=False):
-    configTile = buildConfigTile(tile_id=tile_id, tile_template=tile_template, data=data)
-    if not isTest:
-        res = requests.post(TIPBOARD_URL + "/push", data=configTile)
-        print(f"{res} -> {tile_id}: {res.text}", flush=True)
+             "Developer: 0"
+             ]
+    }
 
 
 def sonde6(isTest):
