+++ src/sensors/sensors9_bigvalue.py
@@ -1,6 +1,6 @@
-import requests, time
-from src.tipboard.app.properties import TIPBOARD_URL
-from src.sensors.utils import getTimeStr, end, buildConfigTile
+import time
+from src.sensors.utils import end
+from src.sensors.utils import sendDataToTipboard
 
 NAME_OF_SENSORS = "big_value"
 TILE_TEMPLATE = "big_value"
@@ -24,16 +24,8 @@ def executeScriptToGetData():
     }
 
 
-def sendDataToTipboard(data=None, tile_template=None, tile_id="", isTest=False):
-    configTile = buildConfigTile(tile_id=tile_id, tile_template=tile_template, data=data)
-    if not isTest:
-        res = requests.post(TIPBOARD_URL + "/push", data=configTile)
-        print(f"{getTimeStr()}:{res} -> {tile_id}: {res.text}", flush=True)
-
-
 def sonde9(isTest):
     start_time = time.time()
     data = executeScriptToGetData()
     sendDataToTipboard(data, tile_template=TILE_TEMPLATE, tile_id=TILE_ID, isTest=isTest)
     end(title=f"sensors9 -> {TILE_ID}", start_time=start_time)
-
